Move findReferenceValidatorScriptByHash to plutus contract
