+++ plutus-contract/changelog.d/20230308_205812_nicolas.biri_plt_967.md
@@ -1,0 +1,4 @@
+### Added
+
+- add a `findReferenceValidatorScripByHash` function to find a `TxOutRef` that
+  references a `TxOut` with the given reference script at a given address.
